requirements.txt
async-generator==1.10
attrs==21.2.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.7
conda==4.3.16
distlib==0.3.3
filelock==3.4.0
h11==0.12.0
idna==3.3
outcome==1.1.0
pipenv==2021.11.15
platformdirs==2.4.0
pycosat==0.6.3
pycparser==2.21
pyOpenSSL==21.0.0
PyYAML==6.0
requests==2.26.0
robotframework==4.1.2
robotframework-pythonlibcore==3.0.0
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==5.1.3
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
wsproto==1.0.0
